+++ packages/api-client/src/CumulusApiClientError.ts
@@ -1,0 +1,8 @@
+export class CumulusApiClientError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = this.constructor.name;
+  }
+}
+
+export default CumulusApiClientError;