+++ packages/api-client/tests/test-executions.js
@@ -1,8 +1,7 @@
 'use strict';
 
 const test = require('ava');
-const rewire = require('rewire');
-const executionsRewire = rewire('../executions');
+const executionsApi = require('../executions');
 
 test.before(async (t) => {
   t.context.testPrefix = 'unitTestStack';
@@ -19,20 +18,17 @@ test('getExecution calls the callback with the expected object', async (t) => {
       path: `/executions/${t.context.arn}`
     }
   };
+
   const callback = async (configObject) => {
-    t.deepEqual(expected, configObject);
+    t.deepEqual(configObject, expected);
     return (t.context.testExecutionReturn);
   };
-  let revertCallback;
-  try {
-    revertCallback = executionsRewire.__set__('invokeApi', callback);
-    await t.notThrowsAsync(executionsRewire.getExecution({
-      prefix: t.context.testPrefix,
-      arn: t.context.arn
-    }));
-  } finally {
-    revertCallback();
-  }
+
+  await t.notThrowsAsync(executionsApi.getExecution({
+    prefix: t.context.testPrefix,
+    arn: t.context.arn,
+    callback
+  }));
 });
 
 test('getExecutions calls the callback with the expected object', async (t) => {
@@ -41,23 +37,42 @@ test('getExecutions calls the callback with the expected object', async (t) => {
     payload: {
       httpMethod: 'GET',
       resource: '/{proxy+}',
-      path: '/executions'
+      path: '/executions',
+      queryStringParameters: undefined
     }
   };
 
   const callback = async (configObject) => {
-    t.deepEqual(expected, configObject);
+    t.deepEqual(configObject, expected);
   };
 
-  let revertCallback;
-  try {
-    revertCallback = executionsRewire.__set__('invokeApi', callback);
-    await t.notThrowsAsync(executionsRewire.getExecutions({
-      prefix: t.context.testPrefix
-    }));
-  } finally {
-    revertCallback();
-  }
+  await t.notThrowsAsync(executionsApi.getExecutions({
+    prefix: t.context.testPrefix,
+    callback
+  }));
+});
+
+test('getExecutions calls the callback with the expected object with query params', async (t) => {
+  const query = { limit: 50 };
+  const expected = {
+    prefix: t.context.testPrefix,
+    payload: {
+      httpMethod: 'GET',
+      resource: '/{proxy+}',
+      path: '/executions',
+      queryStringParameters: query
+    }
+  };
+
+  const callback = async (configObject) => {
+    t.deepEqual(configObject, expected);
+  };
+
+  await t.notThrowsAsync(executionsApi.getExecutions({
+    prefix: t.context.testPrefix,
+    query,
+    callback
+  }));
 });
 
 test('getExecutionStatus calls the callback with the expected object', async (t) => {
@@ -71,16 +86,12 @@ test('getExecutionStatus calls the callback with the expected object', async (t)
   };
 
   const callback = async (configObject) => {
-    t.deepEqual(expected, configObject);
+    t.deepEqual(configObject, expected);
   };
-  let revertCallback;
-  try {
-    revertCallback = executionsRewire.__set__('invokeApi', callback);
-    await t.notThrowsAsync(executionsRewire.getExecutionStatus({
-      prefix: t.context.testPrefix,
-      arn: t.context.arn
-    }));
-  } finally {
-    revertCallback();
-  }
+
+  await t.notThrowsAsync(executionsApi.getExecutionStatus({
+    prefix: t.context.testPrefix,
+    arn: t.context.arn,
+    callback
+  }));
 });